Code cleanup and added test to pytorch.sh
