+++ ci/pytorch.sh
@@ -93,6 +93,7 @@ run_test_config_mgpu(){
         run 3 distributed/test_fusible_ops.py
         run 3 distributed/test_numerics.py
         run 3 distributed/test_torch_fsdp2.py
+        run 3 distributed/test_torch_fsdp2_fp8.py
         run 3 fused_attn/test_fused_attn_with_cp.py
     fi
 }
